move load genesis function into Client
