+++ src/errors.rs
@@ -20,10 +20,10 @@ pub enum ClientError {
     AccountError(AccountError),
     AssetError(AssetError),
     AuthError(FalconError),
-    RpcTypeConversionFailure(ParseError),
     NoteError(NoteError),
     RpcApiError(RpcApiError),
     RpcExpectedFieldMissingFailure(String),
+    RpcTypeConversionFailure(ParseError),
     StoreError(StoreError),
     TransactionExecutionError(TransactionExecutorError),
     TransactionProvingError(TransactionProverError),
